chore(storage): Reorganize shared classes
1. Move `ObjectAccessControl` to `object_metadata.dart`
2. Move `ProjectTeam` to its own file
3. Update tests and imports to reflect this
