+++ packages/google_cloud_storage/lib/google_cloud_storage.dart
@@ -19,5 +19,5 @@ export 'package:google_cloud_rpc/exceptions.dart';
 
 export 'src/bucket_metadata.dart';
 export 'src/client.dart';
-export 'src/object_access_controls.dart';
 export 'src/object_metadata.dart';
+export 'src/project_team.dart';
