+++ packages/google_cloud_storage/lib/src/bucket_metadata.dart
@@ -14,8 +14,8 @@
 
 import 'package:google_cloud_protobuf/protobuf.dart';
 
-import 'object_access_controls.dart';
 import 'object_metadata.dart';
+import 'project_team.dart';
 
 /// The bucket's [Autoclass][] configuration.
 ///
