+++ packages/google_cloud_storage/test/object_metadata_test.dart
@@ -13,10 +13,57 @@
 // limitations under the License.
 
 import 'package:google_cloud_protobuf/protobuf.dart';
-import 'package:google_cloud_storage/src/object_metadata.dart';
+import 'package:google_cloud_storage/google_cloud_storage.dart';
 import 'package:test/test.dart';
 
 void main() {
+  group('ObjectAccessControl', () {
+    test('copyWith', () {
+      final original = ObjectAccessControl(
+        bucket: 'test-bucket',
+        entity: 'user-test@example.com',
+        role: 'READER',
+        projectTeam: ProjectTeam(projectNumber: '123', team: 'editors'),
+        selfLink: Uri.parse('http://example.com'),
+      );
+      final copy = original.copyWith(
+        role: 'OWNER',
+        projectTeam: ProjectTeam(projectNumber: '456', team: 'owners'),
+      );
+
+      expect(copy.bucket, 'test-bucket');
+      expect(copy.entity, 'user-test@example.com');
+      expect(copy.role, 'OWNER');
+      expect(copy.projectTeam?.projectNumber, '456');
+      expect(copy.projectTeam?.team, 'owners');
+      expect(copy.selfLink, Uri.parse('http://example.com'));
+
+      // Original should remain unchanged
+      expect(original.role, 'READER');
+      expect(original.projectTeam?.projectNumber, '123');
+      expect(original.projectTeam?.team, 'editors');
+    });
+
+    test('copyWithout', () {
+      final original = ObjectAccessControl(
+        bucket: 'test-bucket',
+        entity: 'user-test@example.com',
+        role: 'READER',
+        projectTeam: ProjectTeam(projectNumber: '123', team: 'editors'),
+      );
+      final copy = original.copyWithout(projectTeam: true, role: true);
+
+      expect(copy.bucket, 'test-bucket');
+      expect(copy.entity, 'user-test@example.com');
+      expect(copy.role, isNull);
+      expect(copy.projectTeam, isNull);
+
+      // Original should remain unchanged
+      expect(original.role, 'READER');
+      expect(original.projectTeam?.projectNumber, '123');
+    });
+  });
+
   group('ObjectMetadata', () {
     test('copyWith', () {
       final original = ObjectMetadata(
